Move renamer into scope hoisting dir
